Updated getString and getNumer methods to take in two optional parameters.
If $key is not a string, or $min & $max are not numbers, throw an InvalidArgumentException.

If the requested key is missing from the input, throw an OutOfRangeException

If the value is the wrong type, throw a DomainException

If a string is shorter than $min or longer than $max, throw a LengthException

If a number is less than $min or larger than $max, throw a RangeException.